requirements.txt
pytest==8.2.2
selenium==4.22.0
webdriver-manager==4.0.2
python-dotenv==1.0.1

